Rename with-apollo
Name should be same as hoc
