+++ examples/with-apollo/lib/with-apollo.js
