+++ examples/with-apollo/pages/index.js
@@ -2,7 +2,7 @@ import App from '../components/App'
 import Header from '../components/Header'
 import Submit from '../components/Submit'
 import PostList from '../components/PostList'
-import withApolloClient from '../lib/with-apollo-client'
+import withApollo from '../lib/with-apollo'
 
 const IndexPage = props => (
   <App>
@@ -12,4 +12,4 @@ const IndexPage = props => (
   </App>
 )
 
-export default withApolloClient(IndexPage)
+export default withApollo(IndexPage)
